Better StopService to fix CI

+++ FlowTomator.Service/StopService.bat
@@ -1,6 +1,8 @@
 @echo off
 
 net stop FlowTomator
-taskkill /f /im FlowTomator.Service.exe
+
+tasklist /FI "IMAGENAME eq FlowTomator.Service.exe" 2>NUL | find /I /N "FlowTomator.Service.exe">NUL
+if "%ERRORLEVEL%"=="0" taskkill /f /im FlowTomator.Service.exe
 
 exit 0
